Fixes top node deep link w/o setting selector
Issue that this fixes: when a deep link to a top node is specified, e.g. `example.com/orgchart?node=top_node`,  it wasn't linking to this node without the settings="show" and startwith="show" shortcode atts.

Fix: in this case, we just highlight that top node, instead of setting it as the root node.